file extensions changed to jsx and table header moved to table component
