--- right-for-life_front/src/components/SuppliesTableHeader/index.js
@@ -1,3 +1,0 @@
-import SuppliesTableHeader from './SuppliesTableHeader';
-
-export default SuppliesTableHeader;
