+++ right-for-life_front/src/components/SuppliesTableRow/SuppliesTableRow.jsx
